refactor: trip preview logic to hook for testing
