+++ app/src/components/ui/TripPreview.tsx
@@ -1,16 +1,8 @@
-import React, { useState, useEffect } from "react";
 import { View, StyleSheet, Image, TouchableOpacity } from "react-native";
 import { Text } from "react-native-paper";
 
 import { useSession } from "@contexts/SessionContext";
-import {
-  getBookmarks,
-  addBookmark,
-  removeBookmark,
-  countModVerifications,
-  countGpsVerifications,
-  countComments,
-} from "@services/socials-service";
+import { useTripPreviewData } from "@hooks/use-trip-preview-data";
 
 import VotingBar from "@components/VotingBar";
 
@@ -48,47 +40,8 @@ const getImageSource = (mode: TransportationMode) => {
 
 export default function TripPreview({ trip }: { trip: FullTrip }) {
   const { user } = useSession();
-
-  const [bookmarked, setBookmarked] = useState(false);
-  const [modVerifications, setModVerifications] = useState(0);
-  const [gpsVerifications, setGpsVerifications] = useState(0);
-  const [commentCount, setCommentCount] = useState(0);
-
-  useEffect(() => {
-    async function fetchData() {
-      if (!user) return;
-
-      const segmentIds = trip.segments
-        .filter((seg) => !seg.id.startsWith("walk-"))
-        .map((seg) => seg.id);
-
-      const [bookmarks, modCount, gpsCount, commentCountTemp] = await Promise.all([
-        getBookmarks(user.id),
-        countModVerifications(trip.id),
-        countGpsVerifications(segmentIds),
-        countComments(trip.id),
-      ]);
-
-      setBookmarked(bookmarks.includes(trip.id));
-      setModVerifications(modCount);
-      // TODO test GPS verification count
-      setGpsVerifications(gpsCount);
-      setCommentCount(commentCountTemp);
-    }
-
-    fetchData();
-  }, [user, trip]);
-
-  const toggleBookmark = async () => {
-    if (!user) return;
-
-    if (bookmarked) {
-      await removeBookmark(user.id, trip.id);
-    } else {
-      await addBookmark(user.id, trip.id);
-    }
-    setBookmarked(!bookmarked);
-  };
+  const { bookmarked, modVerifications, gpsVerifications, commentCount, toggleBookmark } =
+    useTripPreviewData(user?.id || null, trip);
 
   const totalDuration = trip.segments.reduce((sum, seg) => sum + seg.duration, 0);
   const totalCost = trip.segments.reduce((sum, seg) => sum + seg.cost, 0);
